requirements.txt
transformers>=4.37.0
torch>=2.0.0
num2words
decord
Pillow 